pyproject.toml
[tool.poetry]
name = "sause-demo-ui-autotests"
version = "0.1.0"
description = ""
authors = ["Meerim Sabytkyzy <skmeerim1999@gmail.com>"]
readme = "README.md"
packages = [{include = "sause_demo_ui_autotests"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "7.2.2"
selene = "2.0.0b17"
allure-pytest = "^2.13.0"
python-dotenv = "^0.21.0"



[tool.pytest.ini_options]
addopts="-vv --alluredir=allure-results"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
